refactor: Edit navigation model directory
