+++ app/src/main/java/org/sopt/collaboration/campuspick/feature/main/MainTab.kt
@@ -2,8 +2,8 @@ package org.sopt.collaboration.campuspick.feature.main
 
 import androidx.compose.runtime.Composable
 import org.sopt.collaboration.campuspick.R
-import org.sopt.collaboration.campuspick.core.designsystem.navigation.MainTabRoute
-import org.sopt.collaboration.campuspick.core.designsystem.navigation.Route
+import org.sopt.collaboration.campuspick.core.ui.navigation.MainTabRoute
+import org.sopt.collaboration.campuspick.core.ui.navigation.Route
 
 internal enum class MainTab(
     val iconResId: Int,
